chore: moved user intreface to sep folder
