+++ src/modules/Auth/dto/user.response.dto.ts
@@ -1,4 +1,4 @@
-import { IUser } from '../user'
+import { IUser } from '../types/user.intreface'
 export interface UserResponseDto extends Pick<IUser, 'username'> {
   accessToken: string
   refreshToken: string
